move color.py -> Color.py
