+++ opencsp/common/lib/render_control/RenderControlText.py
@@ -3,7 +3,7 @@
 
 """
 
-import opencsp.common.lib.render.color as cl
+import opencsp.common.lib.render.Color as cl
 
 
 class RenderControlText:
